1.0.2
將數字範圍 由int  擴充到  long

+++ cycle_timing/cycle_timing.cpp
@@ -7,13 +7,15 @@
 
 cycle_timing::cycle_timing(void){}
 
-int cycle_timing::gat_clock() {
+long cycle_timing::gat_clock() {
+
   return ((millis() / 1000) + offset) % m_cycle;
   }
   
 
 	
-void cycle_timing::setup_clock(int cycle,int nowTime){
+void cycle_timing::setup_clock(unsigned long cycle,unsigned long nowTime){
 	m_cycle = cycle;
 	offset = nowTime - (millis() / 1000)% cycle;
+	
 }
